Rbus change for sending messages

+++ src/mqttcm_connect.h
@@ -114,7 +114,7 @@ void init_mqtt_timer (mqtt_timer_t *timer, int max_count);
 void convertToUppercase(char *deviceId);
 int writeToDBFile(char *db_file_path, char *data, size_t size);
 void get_from_file(char *key, char **val, char *filepath);
-void publish_notify_mqtt(char *pub_topic, void *payload, long len);
+void publish_notify_mqtt(char *pub_topic, void *payload, int len);
 int get_global_mqtt_connected();
 void reset_global_mqttConnected();
 void set_global_mqttConnected();
